Move TransitionStatus type
