+++ src/transitions/Transition/index.ts
@@ -1,4 +1,2 @@
-export * from './types';
-
 export * from './Transition.js';
 export { default } from './Transition.js';
