--- src/transitions/Transition/types.ts
@@ -1,3 +1,0 @@
-import { ENTERING, ENTERED, EXITING, EXITED, UNMOUNTED } from './index';
-
-export type TransitionStatus = typeof ENTERING | typeof ENTERED | typeof EXITING | typeof EXITED | typeof UNMOUNTED;
